airbyte-ci: introduce ConnectorContext
